+++ flag/pull_test.go
@@ -8,7 +8,7 @@ import (
 
 func TestPull(t *testing.T) {
 	f := BuildPull(true, true, true)
-	assert.Equal(t, 1, f&PullCommitOffset)
+	assert.Equal(t, int32(1), f&PullCommitOffset)
 	f = ClearCommitOffset(f)
-	assert.Equal(t, 0, f&PullCommitOffset)
+	assert.Equal(t, int32(0), f&PullCommitOffset)
 }
